stock-backtest: 엔진 결함 수정 + Phase 1 누적 미커밋분 (TASK-200~216)
엔진 결함 수정 (2026-04-30 BTC 분기+MA 시나리오에서 발견):
- TASK-211: filter fail 시 보유 청산 누락 (engine.py L219 if target_weights 분기 제거).
  빈 dict 도 정상 결과 — strategy.py L154 contract 위반이었음.
- TASK-212: TradeFill.settlement_date 추가, services/backtest_runner.py 의
  trade.time fallback to now() 제거 → settlement_date(D+1) 정확 반영.
- TASK-213: DataSource.earliest_available() 신규, 자산의 최초 가용일부터 백필.
  yfinance/pykrx 어댑터 구현 + pipeline._resolve_start 재배선.
- TASK-214: yfinance auto_adjust=True (분할/배당 소급 보정 임시처방).
  pykrx 한계 명시 → BLOCKER-003 (corporate_actions Phase 2 정공법).
- TASK-215: 골든 baseline 9 케이스 통합 검증 — fixture 가 순수 math.sin 이라
  재생성 불필요 결정 (12 passed).
- TASK-216: BTC asset_id=56 ohlcv 재백필 — 1825→4242 rows
  (MIN=2014-09-17, MAX=2026-04-28). TASK-213 정책 정상 작동 검증.

선행 누적 미커밋분 (Phase 1 후속):
- TASK-200~205: fractional crypto qty (alembic 0004_fractional_qty),
  AssetWeightMap / FilterConfigBuilder UI 위젯, e2e/persona_first_use 테스트.

+++ projects/stock-backtest/backend/app/api/backtests.py
@@ -187,7 +187,7 @@ def get_backtest_result(
             time=row.time,
             asset_id=row.asset_id,
             side=row.side,  # type: ignore[arg-type]
-            qty=int(row.qty),
+            qty=float(row.qty),
             price=float(row.price),
             commission=float(row.commission),
             currency=row.currency,
